% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RotationForest.R
\name{BuildModel}
\alias{BuildModel}
\title{Builds a single decision tree using rotation forest methodology (Rodriguez et al. 2006)}
\usage{
BuildModel(dependent, response, npredictor, frac = 0.75, ...)
}
\arguments{
\item{dependent}{a data frame of x dependent vectors}

\item{response}{a data frame of y response values}

\item{npredictor}{the number of predictor variables to use in each rotation}

\item{frac}{the fraction of data points to use in each ensemble model}

\item{...}{additional arguments to pass to the rpart function}
}
\value{
A list containing the rpart object and rotation matrix
}
\description{
BuildModel builds one decision tree with a data frame of dependent and response vectors.
It also requires the number of predictor variables to use in each rotation as well as the
fraction of data points to use in each ensemble model (default = 0.75).
}

