% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RotationForest.R
\name{predict.RotationForest}
\alias{predict.RotationForest}
\title{Provides a predict function for an object of class RotationForest}
\usage{
\method{predict}{RotationForest}(RotationForestObject, dependent,
  prob = FALSE)
}
\arguments{
\item{RotationForestObject}{an object of class RotationForest
(returned from the constructor RotationForest(...))}

\item{dependent}{a data frame of the X predictor values}

\item{prob}{A logical indicating whether probabilities of existing in each class are returned
(as opposed to the default predictions)}
}
\value{
A vector of predictions (or a table of probabilities) of the different classes
}
\description{
Predict allows for O(N) prediction based on an object of class RotationForest
where N is the length of the dataframe dependent array.
}
\examples{
data <- read.table("C:/Users/Manan/Desktop/RotationForest/inst/extdata/balance-scale.data", sep = ",", header = FALSE)
data.dependent <- data[,-1]
data.response <- data[,1]
data.response <- as.factor(data.response)
total <- data.frame(data.response, data.dependent)
groups <- sample(rep(1:10, times = ceiling(nrow(total) / 19)), size = nrow(total), replace = TRUE)
data.train <- total[!groups \%in\% 1,]
data.test <- total[groups \%in\% 1,]
fit <- RotationForest(data.train[,-1], data.train[,1], 2, 10)
predict <- predict(fit, data.dependent, prob = FALSE)
}

