% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RotationForest.R
\name{rotationForest}
\alias{rotationForest}
\title{Defines the constructor for the rotationForest module}
\usage{
rotationForest(xdf, ydf, npredictor, ntree = 10, verbose = F, ...)
}
\arguments{
\item{xdf}{a data frame of X dependent vectors}

\item{ydf}{a data frame of Y response values}

\item{npredictor}{the number of predictor variables that are to be used in each rotation}

\item{ntree}{the number of trees that are to be used to train the ensemble}

\item{verbose}{a logical, set true for classification output to be printed}

\item{...}{extra variables to be passed on to the rpart function}
}
\value{
an object of class rotationForest
}
\description{
Requires the dependent and response values (data frames), the number
of predictor variables to use in each rotation, the number of trees to
train, and a logical for determining if progress should be printed
}
\examples{
fpath <- system.file("extdata", "balance-scale.data", package="rotationForest")
data <- read.table(fpath, sep = ",", header = FALSE)
data.dependent <- data[,-1]
data.response <- data[,1]
data.response <- as.factor(data.response)
total <- data.frame(data.response, data.dependent)
groups <- sample(rep(1:10, times = ceiling(nrow(total) / 19)), size = nrow(total), replace = TRUE)
data.train <- total[!groups \%in\% 1,]
data.test <- total[groups \%in\% 1,]
fit <- rotationForest(data.train[,-1], data.train[,1], 2, 10)
}

