% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotationForestNews.R
\name{rotationForestNews}
\alias{rotationForestNews}
\title{Display the NEWS file}
\usage{
rotationForestNews()
}
\description{
\code{rotationForestNews} shows the NEWS file of the rotationForest package.
}
\examples{
rotationForestNews()
}
\author{
Michel Ballings and Dirk Van den Poel, Maintainer: \email{Michel.Ballings@GMail.com}
}
